refactor: one route to map and updated tests
