+++ application/app/_layout.tsx
@@ -4,18 +4,6 @@ export default function RootLayout() {
   return (
     <Stack>
       <Stack.Screen 
-        name="SGW_Map"
-        options={{
-          headerTitle: "SGW map",
-        }}
-      />
-      <Stack.Screen 
-        name="Loyola_Map"
-        options={{
-          headerTitle: "Loyola map",
-        }}
-      />
-      <Stack.Screen 
         name="CampusMap"
         options={{
           headerTitle: "Campus map",
