--- application/components/__tests__/dummy_test.tsx
@@ -1,5 +1,0 @@
-describe('Dummy Test', () => {
-    it('always passes', () => {
-      expect(true).toBe(true);
-    });
-  });
